pyproject.toml
[project]
name = "align-spark-build"
version = "0.1.0"
description = "Build pipeline for align-spark demo"
requires-python = ">=3.10"
dependencies = ["align-utils", "pyyaml"]
